Runs the Offset Measurement Code
Put amount of time to run in time_end, pin number in pin variable. The code will print the IRIGB timestamp alongside the Pi's timestamp with a "|" character dividing them. The IRIG-B decoded value is on the left an the Pi's timestamp value is on the right. The output is saved in a text file name 'Pulse_Reader'