Updated decoder and timestamp code
Uses less steps to decode pulse, keeps track of just the bits it needs to decode the pulse and timestamps more effectively. Improved the way it evaluates the date it reads in to avoid Overflow Errors that stop the program from running. Puts lists as class variables to be able to see what they are keeping track of if there is an error (And makes it easier to get information about the pulse when running it in another file)